fix: consolidate duplicate crypto modules into lib/crypto/encryption.ts
